self-hosted: Journal review fixes
The real catch: the default navigation is what mounts CreatePostButton,
so a shell that replaces it silently cost owners and community members
the way into the composer. JournalShell now mounts it, and that is the
standing contract for any theme shell.

Also from review: the filter state and label logic is extracted into a
shared usePostsFilterState hook consumed by both BlogNavigation and
JournalShell, so no shell can drift from the navigation's behavior; the
masthead action row wraps at narrow viewports instead of overflowing
the measure-width column; light-mode muted text rises to ~4.7:1 AA
over the paper background; and the visibility test walks the section
tree through a typed helper instead of an any cast.
